IndicatorBots/requirements.txt
numpy==1.23.2
pandas==1.5.2
scipy==1.9.3
scikit-learn==1.2.0
requests==2.28.1
python-socketio==5.7.2
eventlet==0.33.3
TA-Lib==0.4.25
dnspython==2.3.0